#1987: changes after codereview : Merged CsvMatcher into CsvSyntaxHighlighter
